Change tests name (from Test* to *Test) because Surefly didn't execute them
